short

+++ 11040A/11040A Autonomous (15-sec).c
@@ -63,7 +63,7 @@ task main()
 
 		float leftD;
 		float rightD;
-		int speed = 100;												//Default Speed, change if needed
+		short speed = 100;												//Default Speed, change if needed
 
 		float mobileGoalDist = 59;							//In inches
